refactor: fragment 패키지 구분 (#43)
